+++ src/main/java/app/controller/MatrixController.java
@@ -12,7 +12,7 @@ public class MatrixController {
    */
   public Matrix sum(Matrix matrix1, Matrix matrix2) throws Error {
     if (matrix1.getWidth() == matrix2.getWidth() && matrix1.getHeight() == matrix2.getHeight()) {
-      float[][] result = new float[matrix1.getWidth()][matrix1.getHeight()];
+      double[][] result = new double[matrix1.getWidth()][matrix1.getHeight()];
       for (int coordX = 0; coordX < matrix1.getWidth(); coordX++) {
         for (int coordY = 0; coordY < matrix1.getHeight(); coordY++) {
           result[coordX][coordY] = matrix1.getData()[coordX][coordY] + matrix2.getData()[coordX][coordY];
@@ -29,7 +29,7 @@ public class MatrixController {
    * @return Matriz resultado de la multiplicación de una matriz con un escalar
    */
   public Matrix multiplicationByScalar(Matrix matrix, int scalar) {
-    float[][] result = new float[matrix.getWidth()][matrix.getHeight()];
+    double[][] result = new double[matrix.getWidth()][matrix.getHeight()];
     for (int coordX = 0; coordX < matrix.getWidth(); coordX++) {
       for (int coordY = 0; coordY < matrix.getHeight(); coordY++) {
         result[coordX][coordY] = scalar * matrix.getData()[coordX][coordY];
@@ -47,7 +47,7 @@ public class MatrixController {
    */
   public Matrix multiplicationOfTwoMatrices(Matrix matrix1, Matrix matrix2) throws Error {
     if (matrix1.getHeight() == matrix2.getWidth()) {
-      float[][] result = new float[matrix1.getWidth()][matrix2.getHeight()];
+      double[][] result = new double[matrix1.getWidth()][matrix2.getHeight()];
       for (int x = 0; x < matrix1.getWidth(); x++) {
         for (int y = 0; y < matrix2.getHeight(); y++) {
           for (int z = 0; z < matrix1.getHeight(); z++) {
@@ -65,7 +65,7 @@ public class MatrixController {
    * @return Matriz inversa
    */
   public Matrix getInverseByGaussJordan(Matrix matrix) {
-    float[][] identity = new float[matrix.getWidth()][matrix.getHeight()];
+    double[][] identity = new double[matrix.getWidth()][matrix.getHeight()];
     for (int i = 0; i < matrix.getWidth(); i++) {
       for (int j = 0; j < matrix.getHeight(); j++) {
         identity[i][j] = i == j ? 1 : 0;
@@ -73,7 +73,7 @@ public class MatrixController {
     }
     for (int y = 0; y < matrix.getHeight(); y++) {
       // matrix.getData()[y][y2] debe ser igual a 1
-      float denominator = matrix.getData().clone()[y][y];
+      double denominator = matrix.getData().clone()[y][y];
       for (int y2 = 0; y2 < matrix.getHeight(); y2++) {
         matrix.getData()[y][y2] /= denominator;
         identity[y][y2] /= denominator;
@@ -81,7 +81,7 @@ public class MatrixController {
       for (int x = 0; x < matrix.getWidth(); x++) {
         if(x != y) {
           // matrix.getData()[x][y] debe ser igual a 0
-          float multiplicator = matrix.getData().clone()[x][y];
+          double multiplicator = matrix.getData().clone()[x][y];
           for (int y2 = 0; y2 < matrix.getHeight(); y2++) {
             matrix.getData()[x][y2] += (-multiplicator) * matrix.getData()[y][y2];
             identity[x][y2] += (-multiplicator) * identity[y][y2];
@@ -101,10 +101,10 @@ public class MatrixController {
    * Ejemplo: ecuacion1 = resultado1, ecuacion2 = resultado2 | matriz = { { ...ecuacion1 }, { ...ecuacion2 } }, valoresIndependientes = { resultado1, resultado2 }
    * @return Arreglo de valores que corresponden con su posición a cada variable del sistema de ecuaciones
    */
-  public float[] getSystemOfEquations(Matrix matrix, float[] independentValues) {
+  public double[] getSystemOfEquations(Matrix matrix, double[] independentValues) {
     for (int y = 0; y < matrix.getHeight(); y++) {
       // matrix.getData()[y][y2] debe ser igual a 1
-      float denominator = matrix.getData().clone()[y][y];
+      double denominator = matrix.getData().clone()[y][y];
       independentValues[y] /= denominator;
       for (int y2 = 0; y2 < matrix.getHeight(); y2++) {
         matrix.getData()[y][y2] /= denominator;
@@ -112,7 +112,7 @@ public class MatrixController {
       for (int x = 0; x < matrix.getWidth(); x++) {
         if(x != y) {
           // matrix.getData()[x][y] debe ser igual a 0
-          float multiplicator = matrix.getData().clone()[x][y];
+          double multiplicator = matrix.getData().clone()[x][y];
           independentValues[x] += (-multiplicator) * independentValues[y];
           for (int y2 = 0; y2 < matrix.getHeight(); y2++) {
             matrix.getData()[x][y2] += (-multiplicator) * matrix.getData()[y][y2];
